Update sample to use Lift 2.2 and GAE/J 1.4.0.
The parent POM file now contains most of the project plugin and dependency settings. The module POM files generally just list the plugins to use and dependencies to include.

The logging library was switched from Log4J to java.util.logging in order to be compatible with GAE/J. SLF4J is a great shim for switching runtime logging dependencies.

I avoided modifying the style or whitespace of the source files to reduce clutter in this changeset. The POM files are an exception.